#28 Testeanbdo endpoints: Añadiendo la descripción del productor al DTO y cambiando altitudMedia a decimal para que coincida con la tabla de la BD.

+++ API/WCA.Application/DTOs/CafeLoteDto.cs
@@ -13,7 +13,7 @@
         public int ProcesoId { get; set; }
         public int VariedadId { get; set; }
         public int TuesteId { get; set; }
-        public double AltitudMedia { get; set; }
+        public decimal AltitudMedia { get; set; }
         public string? DescripcionExtendida { get; set; }
     }
 }
